site conditions displays both map and forms. Logo is now on top. SSURGO location marker fix. error fix
