+++ src/pages/Calculator/Steps/SiteCondition/LocationComponent/MapComponent/mapComponent.css
@@ -8,3 +8,6 @@
 .map-component-dropdown {
     padding-bottom: 20px;
 }
+.map-container div[class*='map_infobar__'] {
+  bottom: 9%;
+}
