+++ src/pages/Calculator/Steps/steps.css
@@ -2,7 +2,7 @@
     margin-bottom: 20px;
 }
 .site-condition-container {
-    padding: 15px 13% !important;
+    padding: 15px 15px !important;
 }
 site-condition-header {
     color: #4f5f30 !important;
